pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Mickael Fioretti <fioretti.mickael@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.6"
pyright = "^1.1.381"

[tool.pyright]
typeCheckingMode = "basic"
reportIncompatibleVariableOverride = false

[tool.ruff] # https://github.com/charliermarsh/ruff#reference
line-length = 100
target-version = "py310"
select = ["ALL"]
ignore = ["D100", "D103"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
